+++ qtarmsim/model/memorydumpproxymodel.py
@@ -21,8 +21,8 @@
 
 import sys
 
-from PySide2 import QtCore, QtGui, QtWidgets
-from PySide2.QtCore import Qt
+from PySide6 import QtCore, QtGui, QtWidgets
+from PySide6.QtCore import Qt
 
 from .common import InputToHex, DataTypes
 from .memorymodel import MemoryModel