+++ qtarmsim/model/simpletreemodel.py
@@ -19,7 +19,7 @@
 ###########################################################################
 
 
-from PySide2 import QtCore
+from PySide6 import QtCore
 
 
 class TreeItem(object):