+++ qtarmsim/utils.py
@@ -16,7 +16,7 @@
 #                                                                         #
 ###########################################################################
 
-from PySide2 import QtGui
+from PySide6 import QtGui
 
 
 def getMonoSpacedFont():