+++ qtarmsim/window/preferencesdialog.py
@@ -17,7 +17,7 @@
 ###########################################################################
 
 
-from PySide2 import QtGui, QtCore, QtWidgets
+from PySide6 import QtGui, QtCore, QtWidgets
 
 from ..ui.preferences import Ui_PreferencesDialog